use multiple shards per index
An empty `settings` dict results in one shard and one replica,
which has led to noticeably unbalanced data distribution on the IL3
opensearch cluster.  I cowboyed a change like this last week (the
copy of pm2os.py on there seems to pre-date the multi-target work)
and each node now has almost the same amount of data on it.